requirements.txt
openai
argparse
python-dotenv